finished payments and every single bug fix

+++ src/app/api/Categories/route.js
@@ -1,18 +1,24 @@
 import { Category } from "@/models/category";
 import mongoose from "mongoose";
-
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const { name } = await req.json();
-  const categoryDoc = await Category.create({ name });
-  return Response.json(categoryDoc);
+  if (await isAdmin()) {
+    const categoryDoc = await Category.create({ name });
+    return Response.json(categoryDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
   const { _id, name } = await req.json();
-  await Category.updateOne({ _id }, { name });
+  if (await isAdmin()) {
+    await Category.updateOne({ _id }, { name });
+  }
   return Response.json(true);
 }
 
@@ -21,10 +27,12 @@ export async function GET() {
   return Response.json(await Category.find());
 }
 
-export async function DELETE(req){
+export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
-  const _id = url.searchParams.get('_id');
-  await Category.deleteOne({_id});
+  const _id = url.searchParams.get("_id");
+  if (await isAdmin()) {
+    await Category.deleteOne({ _id });
+  }
   return Response.json(true);
 }
