+++ src/app/api/menuitems/route.js
@@ -1,31 +1,38 @@
 import mongoose from "mongoose";
 import { MenuItem } from "@/models/menuitem";
+import { isAdmin } from "@/app/api/auth/[...nextauth]/route";
 
-export async function POST(req){
-    mongoose.connect(process.env.MONGO_URL);
-    const data = await req.json();
+export async function POST(req) {
+  mongoose.connect(process.env.MONGO_URL);
+  const data = await req.json();
+  if (await isAdmin()) {
     const menuItemdoc = await MenuItem.create(data);
     return Response.json(menuItemdoc);
-};
+  } else {
+    return Response.json({});
+  }
+}
 
-export async function PUT(req){
-    mongoose.connect(process.env.MONGO_URL);
-    const {_id, ...data} = await req.json();
+export async function PUT(req) {
+  mongoose.connect(process.env.MONGO_URL);
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
     await MenuItem.findByIdAndUpdate(_id, data);
-    return Response.json(true);
-};
+  }
+  return Response.json(true);
+}
 
-export async function GET(){
-    mongoose.connect(process.env.MONGO_URL);
-    return Response.json(
-        await MenuItem.find()
-    );
-};
+export async function GET() {
+  mongoose.connect(process.env.MONGO_URL);
+  return Response.json(await MenuItem.find());
+}
 
-export async function DELETE(req){
-    mongoose.connect(process.env.MONGO_URL);
-    const url = new URL(req.url);
-    const _id = url.searchParams.get('_id');
-    await MenuItem.deleteOne({_id});
-    return Response.json(true);
+export async function DELETE(req) {
+  mongoose.connect(process.env.MONGO_URL);
+  const url = new URL(req.url);
+  const _id = url.searchParams.get("_id");
+  if (await isAdmin()) {
+    await MenuItem.deleteOne({ _id });
+  }
+  return Response.json(true);
 }
